add support for items to include skill-related script
the new SkillScript field in Item DB allows skills that have different
behavior based on some catalyst item to delegate the item-specific
effect to the ItemDB entry.

this is useful for skills like Guillotine Cross poisoning weapon where
the chosen poison defines the effect it will cause in a target.